Generalize using aiter and anext in tests
Provide both functions via the main test module for all supported Python
versions.
